requirements.txt
discord.py
openai
PyYAML